15.05.2022 add function in admin panel photos link

+++ admin/photo/del_photo.php
@@ -1,9 +1,11 @@
 <?php
-$url = mysqli_real_escape_string($link, $_POST['url']);
-$file_path = '../../' . $url;
-unlink($file_path);
-
 include '../../inc/connect_db.php';
+$url = mysqli_real_escape_string($link, $_POST['url']);
+$is_url = filter_var($url, FILTER_VALIDATE_URL);
+if (!$is_url) {
+    $file_path = '../../' . $url;
+    unlink($file_path);
+}
 $sql = "DELETE FROM photo WHERE url='$url'";
 mysqli_query($link, $sql);
 mysqli_close($link);
